requirements.txt
aiofiles           ==23.2.1  
annotated-types    ==0.6.0   
anyio              ==4.2.0   
asgiref            ==3.7.2   
asttokens          ==2.4.1   
bcrypt            == 4.1.2   
certifi           == 2024.2.2
cffi               ==1.16.0  
charset-normalizer ==3.3.2   
click              ==8.1.7   
colorama          == 0.4.6   
confluent-kafka   == 2.3.0   
cryptography      == 42.0.2
decorator         == 5.1.1
distlib           == 0.3.8
dnspython         == 1.16.0
ecdsa            ==  0.18.0
executing        ==  2.0.1
fastapi          ==  0.68.0
filelock         ==  3.13.1
h11              ==  0.14.0
idna             ==  3.6
ipython          ==  8.21.0
itsdangerous     ==  2.1.0
jedi             ==  0.19.1
Jinja2           ==  3.0.1
MarkupSafe       ==  2.1.4
matplotlib-inline == 0.1.6
parso            ==  0.8.3
passlib          ==  1.7.4
pip              ==  23.2.1
platformdirs     ==  4.2.0
prompt-toolkit    == 3.0.43
pure-eval        ==  0.2.2
pyasn1          ==   0.5.1
pycparser        ==  2.21
pydantic         ==  1.10.14
pydantic_core    ==  2.16.1
Pygments        ==   2.17.2
PyJWT           ==   2.8.0
pymongo          ==  3.12.0
python-dotenv    ==  0.19.0
python-jose      ==  3.3.0
python-multipart ==  0.0.6
requests      ==     2.31.0
rsa             ==   4.9
six              ==  1.16.0
sniffio          ==  1.3.0
stack-data       ==  0.6.3
starlette       ==   0.14.2
traitlets        ==  5.14.1
typing_extensions  ==4.9.0
urllib3         ==   2.2.0
uvicorn         ==   0.15.0
virtualenv      ==   20.25.1
wcwidth         ==   0.2.13